requirements.txt
asgiref==3.11.1
Django==4.2.28
djangorestframework==3.16.1
djangorestframework-simplejwt
gunicorn==25.0.2
django-cors-headers
packaging==26.0
psycopg2-binary==2.9.11
requests
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
